Pack XML documentation for public elements
Without adding the XML doc files to the nuget package, the information
isn't usable in IDE tooltips or when checking class metadata and
signature information, which is bad for the developer experience.

This commit adds the XML docs, fixes a few malformed comments and
removes warnings for missing docs on public elements and parameters.